Rename tests/unittest to tests/unit_test
