+++ tests/unit_test/iconscore/typing/test__init__.py
