+++ tests/unit_test/iconscore/typing/test_function.py
