+++ tests/unit_test/iconscore/typing/test_type_hint.py
