+++ tests/unit_test/inv/data/test_value.py
