+++ tests/unit_test/score_loader/test_icon_score_class_loader.py
