+++ tests/unit_test/test_icon_config.py
